requirements.txt
# Mini-JARVIS Dependencies

# Core
python-dotenv>=1.0.0

# HTTP client for Ollama and APIs
httpx>=0.25.0

# Type hints (Python 3.11+ has most, but some extras are useful)
typing-extensions>=4.8.0

# Tools - Knowledge
wikipedia>=1.4.0
arxiv>=2.1.0

# Tools - Search
ddgs>=6.0.0  # DuckDuckGo search (renamed from duckduckgo-search)

# RAG Pipeline (Phase 4)
chromadb>=0.4.0  # Vector database for long-term memory
sentence-transformers>=2.2.0  # Local embeddings (all-MiniLM-L6-v2)
PyPDF2>=3.0.0  # PDF support (optional, for PDF ingestion)

# Testing
pytest>=7.4.0  # Testing framework
pytest-asyncio>=0.21.0  # Async test support
pytest-cov>=4.1.0  # Coverage reporting

